Move deprecate_arguments from circuit.instruction to util.
