+++ qiskit/util.py
@@ -18,6 +18,7 @@ import re
 import socket
 import sys
 import warnings
+import functools
 
 import psutil
 
@@ -95,3 +96,30 @@ def _has_connection(hostname, port):
         return True
     except Exception:  # pylint: disable=broad-except
         return False
+
+
+def deprecate_arguments(kwarg_map):
+    """Decorator to automatically alias deprecated agrument names and warn upon use."""
+    def decorator(func):
+        @functools.wraps(func)
+        def wrapper(*args, **kwargs):
+            if kwargs:
+                _rename_kwargs(func.__name__, kwargs, kwarg_map)
+            return func(*args, **kwargs)
+        return wrapper
+    return decorator
+
+
+def _rename_kwargs(func_name, kwargs, kwarg_map):
+    for old_arg, new_arg in kwarg_map.items():
+        if old_arg in kwargs:
+            if new_arg in kwargs:
+                raise TypeError('{} received both {} and {} (deprecated).'.format(
+                    func_name, new_arg, old_arg))
+
+            warnings.warn('{} keyword argument {} is deprecated and '
+                          'replaced with {}.'.format(
+                              func_name, old_arg, new_arg),
+                          DeprecationWarning, stacklevel=3)
+
+            kwargs[new_arg] = kwargs.pop(old_arg)
